+++ app/src/main/java/app/threedollars/manager/BottomNavItem.kt
@@ -2,7 +2,6 @@ package app.threedollars.manager
 
 sealed class BottomNavItem(var icon : Int, var screenRoute : String) {
     object Home : BottomNavItem(R.drawable.ic_home,"home")
-    object Category: BottomNavItem(R.drawable.ic_category,"category")
-    object AddStore: BottomNavItem(R.drawable.ic_write,"addStore")
+    object StoreManagement: BottomNavItem(R.drawable.ic_truck,"storeManagement")
     object MyPage: BottomNavItem(R.drawable.ic_my,"myPage")
 }
