+++ app/src/main/java/app/threedollars/manager/MainActivity.kt
@@ -1,13 +1,10 @@
 package app.threedollars.manager
 
 import android.os.Bundle
-import android.util.Log
 import androidx.activity.compose.setContent
-import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.Scaffold
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
@@ -15,71 +12,30 @@ import androidx.navigation.compose.rememberNavController
 import app.threedollars.manager.screen.*
 import androidx.compose.ui.tooling.preview.Preview
 import app.threedollars.manager.main.ui.HomeScreen
-import app.threedollars.manager.sign.LoginNavItem
-import app.threedollars.manager.sign.ui.LoginScreen
-import app.threedollars.manager.sign.ui.SignScreen
-import app.threedollars.manager.sign.ui.SplashScreen
-import app.threedollars.manager.sign.ui.WaitingScreen
-import app.threedollars.manager.sign.viewmodel.LoginViewModel
-import com.kakao.sdk.auth.model.OAuthToken
-import com.kakao.sdk.common.model.ClientError
-import com.kakao.sdk.common.model.ClientErrorCause
-import com.kakao.sdk.user.UserApiClient
+import app.threedollars.manager.storeManagement.ui.StoreManagementScreen
 import dagger.hilt.android.AndroidEntryPoint
 
 
 @AndroidEntryPoint
 class MainActivity : AppCompatActivity() {
-
-    private val loginViewModel: LoginViewModel by viewModels()
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
         setContent {
-            LoginNavigationGraph(loginViewModel) {
-                loginKakao()
-            }
+            MainScreenView()
         }
     }
 
-    private fun loginKakao() {
-        val callback: (OAuthToken?, Throwable?) -> Unit = { token, error ->
-            if (error != null) {
-                Log.e("kakao", "카카오계정으로 로그인 실패", error)
-            } else if (token != null) {
-                loginViewModel.login(token.accessToken)
-                Log.i("TAG", "카카오계정으로 로그인 성공 ${token.accessToken}")
-            }
-        }
-
-        if (UserApiClient.instance.isKakaoTalkLoginAvailable(this)) {
-            UserApiClient.instance.loginWithKakaoTalk(this) { token, error ->
-                if (error != null) {
-                    Log.e("kakao", "카카오톡으로 로그인 실패", error)
-                    if (error is ClientError && error.reason == ClientErrorCause.Cancelled) {
-                        return@loginWithKakaoTalk
-                    }
-                    UserApiClient.instance.loginWithKakaoAccount(this, callback = callback)
-                } else if (token != null) {
-                    loginViewModel.login(token.accessToken)
-                    Log.i("TAG", "카카오톡으로 로그인 성공 ${token.accessToken}")
-                }
-            }
-        } else {
-            UserApiClient.instance.loginWithKakaoAccount(this, callback = callback)
-        }
-    }
 }
 
 @Preview
 @Composable
 fun MainScreenView() {
     val navController = rememberNavController()
-//    Scaffold(
-//        bottomBar = { BottomNavigation(navController = navController) }
-//    ) {
-//        NavigationGraph(navController = navController)
-//    }
+    Scaffold(
+        bottomBar = { BottomNavigation(navController = navController) }
+    ) {
+        NavigationGraph(navController = navController)
+    }
 }
 
 @Composable
@@ -88,37 +44,11 @@ fun NavigationGraph(navController: NavHostController) {
         composable(BottomNavItem.Home.screenRoute) {
             HomeScreen()
         }
-        composable(BottomNavItem.Category.screenRoute) {
-            CategoryScreen()
-        }
-        composable(BottomNavItem.AddStore.screenRoute) {
-            AddStoreScreen()
+        composable(BottomNavItem.StoreManagement.screenRoute) {
+            StoreManagementScreen()
         }
         composable(BottomNavItem.MyPage.screenRoute) {
             MyPageScreen()
-        }
-    }
-}
-
-@Composable
-fun LoginNavigationGraph(viewModel: LoginViewModel, loginKakao: () -> Unit) {
-    val navController = rememberNavController()
-
-    NavHost(
-        navController, startDestination = LoginNavItem.Splash.screenRoute, modifier = Modifier.fillMaxSize()
-    ) {
-        composable(LoginNavItem.Splash.screenRoute) {
-            SplashScreen(navController)
-        }
-
-        composable(LoginNavItem.Login.screenRoute) {
-            LoginScreen(navController, loginKakao, viewModel)
-        }
-        composable(LoginNavItem.Sign.screenRoute) {
-            SignScreen(navController)
-        }
-        composable(LoginNavItem.Waiting.screenRoute) {
-            WaitingScreen(navController)
         }
     }
 }
