+++ app/src/main/java/app/threedollars/manager/screen/BottomNaviagtionScreen.kt
@@ -16,8 +16,7 @@ import app.threedollars.manager.R
 fun BottomNavigation(navController: NavController) {
     val items = listOf(
         BottomNavItem.Home,
-        BottomNavItem.Category,
-        BottomNavItem.AddStore,
+        BottomNavItem.StoreManagement,
         BottomNavItem.MyPage
     )
     androidx.compose.material.BottomNavigation(
