+++ app/src/main/java/app/threedollars/manager/sign/LoginNavItem.kt
@@ -5,4 +5,5 @@ sealed class LoginNavItem(var screenRoute: String) {
     object Login : LoginNavItem("login")
     object Sign : LoginNavItem("sign")
     object Waiting : LoginNavItem("waiting")
+    object Home : LoginNavItem("home")
 }
