+++ app/src/main/java/app/threedollars/manager/sign/ui/SplashScreen.kt
@@ -1,5 +1,6 @@
 package app.threedollars.manager.sign.ui
 
+import android.content.Intent
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
@@ -7,20 +8,22 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalContext
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavController
+import app.threedollars.manager.MainActivity
 import app.threedollars.manager.R
 import app.threedollars.manager.sign.LoginNavItem
-import app.threedollars.manager.sign.viewmodel.LoginViewModel
 import app.threedollars.manager.sign.viewmodel.SplashViewModel
+import app.threedollars.manager.util.findActivity
 import com.airbnb.lottie.compose.*
 import kotlinx.coroutines.delay
 
 @Composable
 fun SplashScreen(navController: NavController, viewModel: SplashViewModel = hiltViewModel()) {
     val navItem by viewModel.loginNavItem.collectAsStateWithLifecycle(null)
-
+    val context = LocalContext.current
 
     val composition by rememberLottieComposition(LottieCompositionSpec.RawRes(R.raw.partners_splash))
     val lottieAnimatable = rememberLottieAnimatable()
@@ -45,8 +48,13 @@ fun SplashScreen(navController: NavController, viewModel: SplashViewModel = hilt
     LaunchedEffect(Unit) {
         delay(2000L)
         navItem?.let {
-            navController.popBackStack()
-            navController.navigate(it.screenRoute)
+            if (it == LoginNavItem.Home) {
+                context.startActivity(Intent(context, MainActivity::class.java))
+                context.findActivity().finish()
+            } else {
+                navController.popBackStack()
+                navController.navigate(it.screenRoute)
+            }
         }
     }
 }
